TASK: Fusion performance optimizations
