+++ Neos.Fusion/Classes/Core/RuntimeConfiguration.php
@@ -1,0 +1,181 @@
+<?php
+
+namespace Neos\Fusion\Core;
+
+use Neos\Fusion\Exception;
+use Neos\Flow\Annotations as Flow;
+use Neos\Utility\Arrays;
+
+/**
+ * @Flow\Proxy(false)
+ */
+final class RuntimeConfiguration
+{
+
+    /**
+     * The parsed Fusion configuration
+     *
+     * @var array
+     */
+    private $fusionConfiguration;
+
+    /**
+     * @var array
+     */
+    private $configurationOnPathRuntimeCache = [];
+
+    /**
+     * @var \Closure
+     */
+    private $simpleTypeToArrayClosure;
+
+    /**
+     * @var \Closure
+     */
+    private $shouldOverrideFirstClosure;
+
+    public function __construct(array $fusionConfiguration)
+    {
+        $this->fusionConfiguration = $fusionConfiguration;
+
+        $this->simpleTypeToArrayClosure = function ($simpleType) {
+            return $simpleType === null ? null : [
+                '__eelExpression' => null,
+                '__value' => $simpleType,
+                '__objectType' => null
+            ];
+        };
+
+        $this->shouldOverrideFirstClosure = function ($key, $firstValue, $secondValue): bool {
+            return is_array($secondValue) && isset($secondValue['__stopInheritanceChain']);
+        };
+    }
+
+    /**
+     * Get the expanded Fusion configuration for the given Fusion path
+     *
+     * @param string $fusionPath
+     * @return array
+     * @throws Exception
+     */
+    public function forPath(string $fusionPath): array
+    {
+        if (isset($this->configurationOnPathRuntimeCache[$fusionPath])) {
+            return $this->configurationOnPathRuntimeCache[$fusionPath]['c'];
+        }
+
+        $pathParts = explode('/', $fusionPath);
+        $configuration = $this->fusionConfiguration;
+
+        $pathUntilNow = '';
+        $currentPrototypeDefinitions = [];
+        if (isset($configuration['__prototypes'])) {
+            $currentPrototypeDefinitions = $configuration['__prototypes'];
+        }
+
+        foreach ($pathParts as $pathPart) {
+            $pathUntilNow .= '/' . $pathPart;
+            if (isset($this->configurationOnPathRuntimeCache[$pathUntilNow])) {
+                $configuration = $this->configurationOnPathRuntimeCache[$pathUntilNow]['c'];
+                $currentPrototypeDefinitions = $this->configurationOnPathRuntimeCache[$pathUntilNow]['p'];
+                continue;
+            }
+
+            $configuration = $this->matchCurrentPathPart($pathPart, $configuration, $currentPrototypeDefinitions);
+            $this->configurationOnPathRuntimeCache[$pathUntilNow]['c'] = $configuration;
+            $this->configurationOnPathRuntimeCache[$pathUntilNow]['p'] = $currentPrototypeDefinitions;
+        }
+
+        return $configuration;
+    }
+
+    /**
+     * Matches the current path segment and prepares the configuration.
+     *
+     * @param string $pathPart
+     * @param array $previousConfiguration
+     * @param array $currentPrototypeDefinitions
+     * @return array
+     * @throws Exception
+     */
+    protected function matchCurrentPathPart($pathPart, $previousConfiguration, &$currentPrototypeDefinitions)
+    {
+        if (preg_match('#^([^<]*)(<(.*?)>)?$#', $pathPart, $matches) !== 1) {
+            throw new Exception('Path Part ' . $pathPart . ' not well-formed', 1332494645);
+        }
+
+        $currentPathSegment = $matches[1];
+        $configuration = [];
+
+        if (isset($previousConfiguration[$currentPathSegment])) {
+            $configuration = is_array($previousConfiguration[$currentPathSegment]) ? $previousConfiguration[$currentPathSegment] : $this->simpleTypeToArrayClosure->__invoke($previousConfiguration[$currentPathSegment]);
+        }
+
+        if (isset($configuration['__prototypes'])) {
+            $currentPrototypeDefinitions = Arrays::arrayMergeRecursiveOverruleWithCallback($currentPrototypeDefinitions, $configuration['__prototypes'], $this->simpleTypeToArrayClosure, $this->shouldOverrideFirstClosure);
+        }
+
+        $currentPathSegmentType = null;
+        if (isset($configuration['__objectType'])) {
+            $currentPathSegmentType = $configuration['__objectType'];
+        }
+        if (isset($matches[3])) {
+            $currentPathSegmentType = $matches[3];
+        }
+
+        if ($currentPathSegmentType !== null) {
+            $configuration['__objectType'] = $currentPathSegmentType;
+            $configuration = $this->mergePrototypesWithConfigurationForPathSegment($configuration, $currentPrototypeDefinitions);
+        }
+
+        if (is_array($configuration) && !isset($configuration['__value']) && !isset($configuration['__eelExpression']) && !isset($configuration['__meta']['class']) && !isset($configuration['__objectType']) && isset($configuration['__meta']['process'])) {
+            $configuration['__value'] = '';
+        }
+
+        return $configuration;
+    }
+
+    /**
+     * Merges the prototype chain into the configuration.
+     *
+     * @param array $configuration
+     * @param array $currentPrototypeDefinitions
+     * @return array
+     * @throws Exception
+     */
+    protected function mergePrototypesWithConfigurationForPathSegment($configuration, &$currentPrototypeDefinitions)
+    {
+        $currentPathSegmentType = $configuration['__objectType'];
+
+        if (isset($currentPrototypeDefinitions[$currentPathSegmentType])) {
+            $prototypeMergingOrder = [$currentPathSegmentType];
+            if (isset($currentPrototypeDefinitions[$currentPathSegmentType]['__prototypeChain'])) {
+                $prototypeMergingOrder = array_merge($currentPrototypeDefinitions[$currentPathSegmentType]['__prototypeChain'], $prototypeMergingOrder);
+            }
+
+            $currentPrototypeWithInheritanceTakenIntoAccount = [];
+
+            foreach ($prototypeMergingOrder as $prototypeName) {
+                if (!array_key_exists($prototypeName, $currentPrototypeDefinitions)) {
+                    throw new Exception(sprintf(
+                        'The Fusion object `%s` which you tried to inherit from does not exist. Maybe you have a typo on the right hand side of your inheritance statement for `%s`.',
+                        $prototypeName, $currentPathSegmentType), 1427134340);
+                }
+
+                $currentPrototypeWithInheritanceTakenIntoAccount = Arrays::arrayMergeRecursiveOverruleWithCallback($currentPrototypeWithInheritanceTakenIntoAccount, $currentPrototypeDefinitions[$prototypeName], $this->simpleTypeToArrayClosure, $this->shouldOverrideFirstClosure);
+            }
+
+            // We merge the already flattened prototype with the current configuration (in that order),
+            // to make sure that the current configuration (not being defined in the prototype) wins.
+            $configuration = Arrays::arrayMergeRecursiveOverruleWithCallback($currentPrototypeWithInheritanceTakenIntoAccount, $configuration, $this->simpleTypeToArrayClosure, $this->shouldOverrideFirstClosure);
+
+            // If context-dependent prototypes are set (such as prototype("foo").prototype("baz")),
+            // we update the current prototype definitions.
+            if (isset($currentPrototypeWithInheritanceTakenIntoAccount['__prototypes'])) {
+                $currentPrototypeDefinitions = Arrays::arrayMergeRecursiveOverruleWithCallback($currentPrototypeDefinitions, $currentPrototypeWithInheritanceTakenIntoAccount['__prototypes'], $this->simpleTypeToArrayClosure, $this->shouldOverrideFirstClosure);
+            }
+        }
+
+        return $configuration;
+    }
+}
